gl: Fix compute invocation counts for format handling code

+++ rpcs3/Emu/RSX/GL/GLTexture.cpp
@@ -856,7 +856,7 @@ namespace gl
 				if (driver_caps.ARB_compute_shader_supported)
 				{
 					// 0. Preconf
-					mem_layout.alignment = caps.alignment;
+					mem_layout.alignment = static_cast<u8>(caps.alignment);
 					mem_layout.swap_bytes = op.require_swap;
 					mem_layout.format = gl_format;
 					mem_layout.type = gl_type;
